+++ lib/src/controller.dart
@@ -2,10 +2,21 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 
+class NotInCacheException implements Exception {
+  String toString() => 'The item is not in the cache.';
+}
+
 /// The [CacheController] creates [CacheUpdate]s to inform other classes about
 /// its state of fetching data.
 @immutable
 class CacheUpdate<T> {
+  CacheUpdate({
+    @required this.isFetching,
+    this.data,
+    this.error,
+    this.stackTrace,
+  }) : assert(isFetching != null);
+
   /// Whether the fetching of the original data source is still in progress.
   final bool isFetching;
 
@@ -13,12 +24,10 @@ class CacheUpdate<T> {
   final T data;
   bool get hasData => data != null;
 
-  /// An error that the source threw.
+  /// An error that the source threw and its stack trace.
   final dynamic error;
+  final StackTrace stackTrace;
   bool get hasError => error != null;
-
-  CacheUpdate({@required this.isFetching, this.data, this.error})
-      : assert(isFetching != null);
 }
 
 /// A class that manages fetching data using the provided [fetcher] and saves
@@ -26,23 +35,7 @@ class CacheUpdate<T> {
 /// [fetch] on this class, you can start the process of fetching data
 /// simultaneously from the cache and the original source. To get updates about
 /// this process, you can listen to the [updates] stream.
-/// Call [dispose] after you're done using the [CacheController].
 class CacheController<T> {
-  final Future<T> Function() fetcher;
-  Future<void> _currentFetcher;
-
-  final Future<void> Function(T data) saveToCache;
-  final Future<T> Function() loadFromCache;
-
-  final _updates = StreamController<CacheUpdate<T>>.broadcast();
-  Stream<CacheUpdate<T>> get updates => _updates.stream;
-
-  T _lastData;
-  T get lastData => _lastData;
-
-  dynamic _lastError;
-  dynamic get lastError => _lastError;
-
   CacheController({
     @required this.fetcher,
     @required this.saveToCache,
@@ -51,8 +44,57 @@ class CacheController<T> {
         assert(saveToCache != null),
         assert(loadFromCache != null);
 
-  /// Disposes the internally used stream controller.
-  void dispose() => _updates.close();
+  final Future<T> Function() fetcher;
+  Future<void> _currentFetcher;
+  bool get isFetching => _currentFetcher != null;
+
+  final Future<void> Function(T data) saveToCache;
+  final Future<T> Function() loadFromCache;
+
+  T _lastData;
+  T get lastData => _lastData;
+
+  dynamic _lastError;
+  StackTrace _lastStackTrace;
+  dynamic get lastError => _lastError;
+  StackTrace get lastStackTrace => _lastStackTrace;
+
+  void _addUpdate() {
+    for (final controller in _controllers) {
+      controller.add(CacheUpdate(
+        isFetching: isFetching,
+        data: lastData,
+        error: lastError,
+        stackTrace: lastStackTrace,
+      ));
+    }
+  }
+
+  final Set<StreamController<CacheUpdate<T>>> _controllers = {};
+  Stream<CacheUpdate<T>> get updates {
+    StreamController<CacheUpdate<T>> controller;
+    controller = StreamController<CacheUpdate<T>>(
+      onCancel: () {
+        controller.close();
+        _controllers.remove(controller);
+      },
+    );
+    _controllers.add(controller);
+
+    controller.add(CacheUpdate(
+      isFetching: isFetching,
+      data: lastData,
+      error: lastError,
+      stackTrace: lastStackTrace,
+    ));
+
+    return controller.stream;
+  }
+
+  @Deprecated(
+      "You don't need to call dispose anymore. The CacheController takes care "
+      "of that by itself now.")
+  void dispose() {}
 
   /// Fetches data from the cache and the [fetcher] simultaneously.
   Future<void> fetch() {
@@ -61,14 +103,6 @@ class CacheController<T> {
   }
 
   Future<void> _actuallyFetchData() async {
-    bool fetchingCompleted = false;
-
-    _updates.add(CacheUpdate(
-      isFetching: true,
-      data: _lastData,
-      error: _lastError,
-    ));
-
     // Simultaneously get data from the (probably faster) cache and the
     // original source.
     await Future.wait([
@@ -81,17 +115,10 @@ class CacheController<T> {
             // to do anything. Otherwise, we push a [CacheUpdate] with the
             // [cachedData] so that it can be displayed to the user while the
             // original source is still loading.
-            if (!fetchingCompleted) {
-              _updates.add(CacheUpdate(
-                isFetching: true,
-                data: _lastData,
-                error: _lastError,
-              ));
-            }
-          } catch (error) {
-            // The [loadFromCache] function throwing means that the cache
-            // doesn't contain any cached data yet. In that case, we just fail
-            // silently as the original data will be returned soon.
+            if (isFetching) _addUpdate();
+          } on NotInCacheException {
+            // We just fail silently as the original data will be returned
+            // soon.
           }
         }),
       // Get data from the original source.
@@ -99,19 +126,16 @@ class CacheController<T> {
         try {
           T data = await fetcher();
           _lastError = null;
-          _updates.add(CacheUpdate(isFetching: false, data: data));
+          _lastStackTrace = null;
           saveToCache(data);
-        } catch (error) {
+        } catch (error, stackTrace) {
           _lastError = error;
-          _updates.add(CacheUpdate(
-            isFetching: false,
-            data: _lastData,
-            error: error,
-          ));
+          _lastStackTrace = stackTrace;
+        } finally {
+          _currentFetcher = null;
+          _addUpdate();
         }
       }),
     ]);
-
-    _currentFetcher = null;
   }
 }